+++ build.ps1
@@ -1,6 +1,6 @@
 [cmdletbinding(DefaultParameterSetName = 'task')]
 param(
-    [parameter(ParameterSetName = 'task')]
+    [parameter(ParameterSetName = 'task',Position = 0)]
     [string[]]$Task = 'default',
 
     [parameter(ParameterSetName = 'help')]